Add modules & date feature
