+++ index.css
@@ -25,24 +25,24 @@ nav {
   /* padding: 10px 32px; */
 }
 
-ul {
+nav ul {
   list-style: none;
   display: flex;
   gap: 24px;
 }
 
-li {
+nav li {
   cursor: pointer;
   padding-right: 24px;
   border-right: 3px solid #000;
 }
 
-li:last-child {
+nav li:last-child {
   padding-right: 0;
   border-right: 0 solid transparent;
 }
 
-li.active {
+nav li.active {
   color: blue;
 }
 
@@ -134,4 +134,25 @@ button {
   border: 2px solid #000;
   width: 100px;
   float: right;
+}
+
+.date {
+  padding-top :1.4rem;
+  float: right;
+  font-weight: bold;
+}
+
+.section-contact .container {
+  max-width: 50rem;
+}
+
+footer {
+
+  border: 3px solid black;
+}
+
+.footer{
+  position: fixed;
+  bottom: 0;
+  width: 100%;
 }
