+++ package.json
@@ -4,5 +4,8 @@
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.27.5"
+  },
+  "dependencies": {
+    "luxon": "^3.3.0"
   }
 }
